requirements.txt
click==6.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.2
pytest-flask==0.10.0
python-status==1.0.1
PyYAML==3.11
schematics==1.1.1
six==1.10.0
SQLAlchemy==1.0.14
Werkzeug==0.11.10
